Add Account controller
