+++ app/views/Main.scala.html
@@ -57,11 +57,11 @@
 			</ul>
 			@if(requestContext.user.isDefined) {
 			<ul class="nav navbar-nav navbar-right">
-				<li><a href='@routes.Application.profile()'>Eingeloggt als @requestContext.user.get.name</a></li>
-				<li><a href="@routes.Application.logout()">Logout</a></li>
+				<li><a href='@routes.Account.profile()'>Eingeloggt als @requestContext.user.get.name</a></li>
+				<li><a href="@routes.Account.logout()">Logout</a></li>
 			</ul>
 			} else {
-			<form class="navbar-form navbar-right" method="POST" action='@routes.Application.login()'>
+			<form class="navbar-form navbar-right" method="POST" action='@routes.Account.login()'>
                     		<div class="form-group">
                         		<input type="text" class="form-control logininput" name="username" placeholder="Username" />
                     		</div>
