+++ app/views/Profile.scala.html
@@ -5,7 +5,7 @@
 	<h1>Nutzereinstellungen</h1>
 
 	<h2>Account-Informationen</h2>
-	<form method="POST" action='@routes.Application.profilePost'>
+	<form method="POST" action='@routes.Account.profilePost'>
 		<table class="table">
 			<tbody>
 				<tr>
@@ -35,7 +35,7 @@
 	</form>
 
 	<h2>Passwort ändern</h2>
-	<form method="POST" action='@routes.Application.profilePost'>
+	<form method="POST" action='@routes.Account.profilePost'>
 		<div class="row">
 			<div class="form-group col-md-4">
 				<label for="pw1">Neues Passwort</label><br />
