+++ conf/routes
@@ -2,10 +2,11 @@ GET	/webjars/*file				controllers.WebJarAssets.at(file)
 GET	/assets/*file				controllers.Assets.at(path="/public", file)
 
 GET	/					controllers.Application.index()
-POST	/login					controllers.Application.login()
-GET	/logout					controllers.Application.logout()
-GET	/profile				controllers.Application.profile()
-POST	/profile				controllers.Application.profilePost()
+
+POST	/login					controllers.Account.login()
+GET	/logout					controllers.Account.logout()
+GET	/profile				controllers.Account.profile()
+POST	/profile				controllers.Account.profilePost()
 
 GET	/about/allgemeines			controllers.Application.aboutAllgemeines()
 GET	/about/fachschaften			controllers.Application.aboutFachschaften()
